% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CBMC8K_example_data}
\alias{CBMC8K_example_data}
\title{CBMC 8K example dataset}
\format{
A data frame with 132 rows and 12 variables
}
\usage{
data(CBMC8K_example_data)
}
\description{
A dataset containing expression informations of 12 genes in 11 celltypes. Original dataset from seurat-data package.
}
\details{
\itemize{
\item features.plot. Gene symbol
\item id. Cell type
\item RNA.pct.exp. RNA level : Percentage of cells expressing the gene in the cell type
\item RNA.avg.exp. RNA level : Average expression of the gene in the cell type
\item RNA.avg.exp.log2p1. RNA.avg.exp with log2(x+1) tranformation
\item RNA.avg.exp.scaled. RNA level : Scaled average expression
\item ADT.pct.exp.sup0. ADT level : Percentage of cells expressing the gene in the cell type (with expression >0)
\item ADT.pct.exp.sup.cutoff. ADT level : Percentage of cells expressing the gene in the cell type (with expression > background)
\item ADT.avg.exp. ADT level : Average expression of the gene in the cell type
\item ADT.avg.exp.log2p1. ADT.avg.exp with log2(x+1) tranformation
\item ADT.avg.exp.scaled. ADT level : Scaled average expression
\item canonical_marker. If the gene is a canonical marker of the cell type (yes or no)
}
}
\references{
\url{https://github.com/satijalab/seurat-data}
}
\keyword{datasets}
