% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_plot.R
\encoding{UTF-8}
\name{dot_plot}
\alias{dot_plot}
\title{Dot-plot - Pacman-plot function}
\usage{
dot_plot(data.to.plot, size_var = NA, col_var = NA, text_var = NA,
  shape_var = 16, size_legend = "", col_legend = "",
  shape_legend = "", cols.use = "default", shape.scale = 12,
  text.size = NA, shape_use = "default", scale.by = "radius",
  scale.min = NA, scale.max = NA, plot.legend = TRUE,
  do.return = FALSE, x.lab.rot = TRUE, x.lab.pos = c("both", "top",
  "bottom", "none"), y.lab.pos = c("left", "right", "both", "none"),
  vertical_coloring = NA, horizontal_coloring = NA,
  size.breaks.number = 4, color.breaks.number = 5,
  shape.breaks.number = 5, transpose = FALSE, dend_x_var = NULL,
  dend_y_var = NULL, dist_method = c("euclidean", "maximum",
  "manhattan", "canberra", "binary", "minkowski"),
  hclust_method = c("ward.D", "single", "complete", "average",
  "mcquitty", "median", "centroid", "ward.D2"))
}
\arguments{
\item{data.to.plot}{Input data. Can be a list or a data.frame.
If data.frame : Column 1 = x axis (Factor); Col2= y axis (Factor).
If list : x and y axis are fixed by row and col names of list elements.}

\item{size_var}{If numeric : Column/List index which control shape sizes. This column/element has to be numeric.
Can also be a column/element name or a vector of the same size than the input dataset.
Set to NA if you don't want to control shape size.}

\item{col_var}{If numeric : Column/List index which control shape colors. 
Can also be a column/element name or a vector of the same size than the input dataset.
Set to NA if you don't want to control shape color.}

\item{text_var}{If numeric : Column/List index which control text to add on shapes. 
Can also be a column/element name or a vector of the same size than the input dataset.
Set to NA if you don't want to add text.}

\item{shape_var}{If numeric = Similar to pch : square=15; circle=16; triangle=17. Can also be a column/element name or a vector of the same size than the input dataset.}

\item{size_legend}{Custom name of shape legend.}

\item{col_legend}{Custom name of shape color.}

\item{shape_legend}{Name of the shape legend if shape_var is a vector.}

\item{cols.use}{1 color or a vector containing multiple colors to color shapes.
If coloring is continuous, default colors are taken from a "lightgrey" to "blue" gradient.
If coloring is discrete, default colors are taken from the default ggplot2 palette.}

\item{shape.scale}{Scale the size of the shapes, similar to cex.}

\item{text.size}{Size of text to display on the shapes.}

\item{shape_use}{Shapes to uses (only when shape is controled by a discrete factor). Default shapes : \\u25A0 \\u25CF \\u25C6 \\u2BC8 \\u2BC7 \\u2BC6 \\u2BC5 \\u25D8 \\u25D9 \\u2726  \\u2605 \\u2736 \\u2737.}

\item{scale.by}{Scale the size by size or radius.}

\item{plot.legend}{Plot the legends ?}

\item{do.return}{Return ggplot2 object ?}

\item{x.lab.rot}{Rotate x-axis labels ?}

\item{x.lab.pos}{Where to display x axis labels. This must be one of "bottom","top","both" or "none".}

\item{y.lab.pos}{Where to display y axis labels. This must be one of "left","right","both"or "none".}

\item{vertical_coloring}{Which color use to color the plot vertically ? (colors are repeated untill the end of the plot). Setting vertical and horizontal coloring at the same time is not recommended !}

\item{horizontal_coloring}{Which color use to color the plot horizontally ? (colors are repeated untill the end of the plot). Setting vertical and horizontal coloring at the same time is not recommended !}

\item{size.breaks.number}{Number of shapes with different size to display in the legend.}

\item{color.breaks.number}{Number of labels for the color gradient legend.}

\item{shape.breaks.number}{Number of shapes to display in the legend. Used when shape is controled by a continuous factor only.}

\item{transpose}{Reverse x axis and y axis ?}

\item{dend_x_var}{A vector containing Column/List indexes or Column/List names to compute the x axis dendrogramm.}

\item{dend_y_var}{A vector containing Column/List indexes or Column/List names to compute the y axis dendrogramm.}

\item{dist_method}{The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".}

\item{hclust_method}{The agglomeration method to be used. This must be one of "single", "complete", "average", "mcquitty", "ward.D", "ward.D2", "centroid" or "median".}

\item{scale.min/scale.max}{Set lower and upper limits for scaling, use NA for default values.}
}
\value{
Print the plot and return the ggplot object if do.return=TRUE
}
\description{
Create dotplots to represent two discrete factors (x & y) described by several other factors. Each combination of the two discrete factors (x & y) can be described with : 1 continuous factor (setting shape size), 3 continuous or discrete factors (setting shape type, shape color and text on shape).
}
\author{
Simon Leonard - simon_leonard[a]hotmail.fr
}
