% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PBMC3K_example_data}
\alias{PBMC3K_example_data}
\title{PBMC 3K example dataset}
\format{
A data frame with 312 rows and 13 variables
}
\usage{
data(PBMC3K_example_data)
}
\description{
A dataset containing expression informations of 39 genes in 8 celltypes. Original dataset from seurat-data package.
}
\details{
\itemize{
\item genes.plot. Gene symbol
\item id. Cell type
\item pct.exp. Percentage of cells expressing the gene in the cell type
\item pct.exp2. pct.exp with 2 decimal places
\item pct.exp100. pct.exp.2 x 100
\item avg.exp. Average expression of the gene in the cell type
\item avg.exp.scale. Scaled average expression
\item abs_avg_exp_scale Absolute value of avg.exp.scale
\item avg_logFC. log fold-chage of the average expression of the gene between the cell type and the others
\item avg_logFC2. avg_logFC with 2 decimal places
\item p_val_adj. Adjusted p-value based on bonferroni correction
\item p_val_adj2. p_val_adj with 2 decimal places
\item pval_symb. Adjusted p-value classification ("<1e-100" or "<1e-50" or  "<1e-10" or  "<0.01" or   ">0.01")
}
}
\references{
\url{https://github.com/satijalab/seurat-data}
}
\keyword{datasets}
