% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_dot_plot_dendrogram.R
\encoding{UTF-8}
\name{rotate_dot_plot_dendrogram}
\alias{rotate_dot_plot_dendrogram}
\title{Interactively rotate dendrograms from dot_plot outputs}
\usage{
rotate_dot_plot_dendrogram(dot_plot_output, axis_to_rotate = c("x", "y"))
}
\arguments{
\item{dot_plot_output}{Output from dot_plot function}

\item{axis_to_rotate}{Dendrogram to rotate "x" or "y"}
}
\value{
Print and return rotated dot plot output
}
\description{
Take a output from dot_plot function and allow interactive dendrogram rotation with dendextend package
}
\examples{
# Perform dot_plot
dot_plot_output = dot_plot(...)
# y rotation
r1=rotate_dot_plot_dendrogram(dot_plot_output, "y")
# add x rotation to previous result
r2=rotate_dot_plot_dendrogram(r1, "x")
}
\author{
Simon Leonard - simon_leonard[a]hotmail.fr
}
