% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny_dot_plot.R
\name{Shiny_dot_plot}
\alias{Shiny_dot_plot}
\title{Shiny dotplot}
\usage{
Shiny_dot_plot(RStudio = F)
}
\description{
Shiny application to perform dot-plot pacman-plot
}
\examples{
Shiny_dot_plot()
}
\author{
Simon Leonard - simon_leonard[a]hotmail.fr
}
